# openapi.yaml
openapi: 3.0.3
info:
  title: URL Expander API - Auth Foundation
  version: 1.0.0
  description: |
    Authentication endpoints for account-based access.
    This specification covers auth only (registration, login, token refresh, logout).
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Authentication and session management

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                invalid_name:
                  summary: Name validation
                  value:
                    field: name
                    message: Name must be between 2 and 255 characters
                invalid_email:
                  summary: Email validation
                  value:
                    field: email
                    message: Please provide a valid email address
                invalid_password:
                  summary: Password validation
                  value:
                    field: password
                    message: Password must be at least 8 characters long
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
              example:
                message: An account with this email already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user credentials
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
              example:
                message: Too many login attempts. Please try again later.

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new token pair
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
              example:
                message: Session expired. Please login again

  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token
      operationId: logoutUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '204':
          description: Refresh token revoked
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
              example:
                message: Session expired. Please login again

  /auth/me:
    get:
      tags: [Auth]
      summary: Retrieve authenticated user profile
      operationId: me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          description: Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
              example:
                message: Session expired. Please login again

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthTokens:
      type: object
      required: [access_token, refresh_token, expires_in, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
        token_type:
          type: string
          example: Bearer

    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string

    Me:
      type: object
      required: [user_id, email, name]
      properties:
        user_id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        name:
          type: string

  responses:
    AuthSuccess:
      description: Authentication successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthTokens'

    InvalidCredentials:
      description: Email or password incorrect
      content:
        application/json:
          schema:
            type: object
            required: [message]
            properties:
              message:
                type: string
                example: Email or password is incorrect